requirements.txt
# Phase 1-2 dependencies
discord.py>=2.3.0
pytest>=7.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
